datastore: Consolidate all IndexStore methods
An idea I'm experimenting with.
