+++ datastore/postgres/queries/indexer/delete_layers.sql
@@ -1,0 +1,10 @@
+DELETE FROM
+    layer
+WHERE
+    NOT EXISTS (
+        SELECT
+        FROM
+            manifest_layer
+        WHERE
+            manifest_layer.layer_id = layer.id
+    );
