+++ datastore/postgres/queries/indexer/delete_manifest.sql
@@ -1,0 +1,4 @@
+DELETE FROM
+    manifest
+WHERE
+    hash = ANY($1::TEXT[]) RETURNING manifest.hash;
