+++ datastore/postgres/queries/indexer/insert_layer.sql
@@ -1,0 +1,4 @@
+INSERT INTO
+    layer (hash)
+VALUES
+    ($1) ON CONFLICT DO NOTHING;
