+++ datastore/postgres/queries/indexer/insert_manifest.sql
@@ -1,0 +1,4 @@
+INSERT INTO
+    manifest (hash)
+VALUES
+    ($1) ON CONFLICT DO NOTHING
