+++ datastore/postgres/queries/indexer/insert_manifest_scanned.sql
@@ -1,0 +1,20 @@
+WITH manifests AS (
+    SELECT
+        id AS manifest_id
+    FROM
+        manifest
+    WHERE
+        hash = $1
+)
+INSERT INTO
+    scanned_manifest (manifest_id, scanner_id)
+VALUES
+    (
+        (
+            SELECT
+                manifest_id
+            FROM
+                manifests
+        ),
+        $2
+    );
