+++ datastore/postgres/queries/indexer/insert_scanner.sql
@@ -1,0 +1,4 @@
+INSERT INTO
+    scanner (name, version, kind)
+VALUES
+    ($1, $2, $3) ON CONFLICT (name, version, kind) DO NOTHING;
