+++ datastore/postgres/queries/indexer/scanner_exists.sql
@@ -1,0 +1,11 @@
+SELECT
+    EXISTS(
+        SELECT
+            1
+        FROM
+            scanner
+        WHERE
+            name = $1
+            AND version = $2
+            AND kind = $3
+    );
