+++ datastore/postgres/queries/indexer/select_index_report.sql
@@ -1,0 +1,7 @@
+SELECT
+    scan_result
+FROM
+    indexreport
+    JOIN manifest ON manifest.hash = $1
+WHERE
+    indexreport.manifest_id = manifest.id;
