+++ datastore/postgres/queries/indexer/select_packages.sql
@@ -1,0 +1,13 @@
+SELECT
+    id,
+    name,
+    version,
+    kind,
+    norm_kind,
+    norm_version,
+    module,
+    arch
+FROM
+    package
+WHERE
+    name = $1;
