+++ datastore/postgres/queries/indexer/select_repository.sql
@@ -1,0 +1,8 @@
+SELECT
+    id
+FROM
+    repo
+WHERE
+    name = $1
+    AND key = $2
+    AND uri = $3;
