+++ datastore/postgres/queries/indexer/select_scanner.sql
@@ -1,0 +1,8 @@
+SELECT
+	id
+FROM
+	scanner
+WHERE
+	name = $1
+	AND version = $2
+	AND kind = $3;
